Fix issue where tilespecs could display with the wrong apha value
This was triggered by opening a map or reloading graphics with the Special tab open.
Selecting the Special tab causes the editor to rebuild the render list, which updates alpha values. However, if tilegraphics are in the middle of loading when this occurs, they won't yet be in the render list, and so will not be updated.
The alpha value is now calculated when the tilegraphic *finishes* loading, which prevents this problem.