Ensure palette layer buttons container is not focusable in electron
When experimental web platform features are enabled, the blink
KeyboardFocusableScrollers feature is enabled.
This causes scrollable container to become keyboard-focusable.
See: https://bugs.chromium.org/p/chromium/issues/detail?id=907284

Even trickier, setting tabindex to -1 on a web component causes
everything inside of it to become untabbable.
See: https://github.com/WICG/webcomponents/issues/774

So the solution in this case is to wrap up the action-group component
with a plain old div with tabIndex -1, and let that div handle the
overflow/scrolling instead.