Prevent pointer capture from being released prematurely
The current tool now controls the state of the pointer capture.
The pointer is only captured when the `PointerDownState` changes from
`None`, and only released when it changes back to `None`.
This prevents issues where, for example, we capture the pointer due to a
left click and prematurely release it due to a right click.