Handle texture manager destruction when loading texture cache entries
This can happen when rapidly refreshing graphics. The phaser `Game` is
destroyed while the GFX is being downloaded, which causes the texture
manager and the underlying texture cache multitexture pages to be
destroyed right before the texture cache tries to draw the new cache
entry to one of those pages.