Pass GFX loader directly to editor and palette scene constructors
Previous these scenes were accessing the GFX loader through a
combination of the Phaser data manager and the magic component data
forwarding in the `PhaserInstance` component.

In both cases this was undesirable, as the GFX loader and scene
lifetimes are strongly coupled. The moment we get a new GFX loader, we
construct a new scene to use **that** GFX loader instead.

Previously, it was possible for a scene to briefly use the wrong GFX
loader due to this event-based component data forwarding indirection.
This was easy to trigger by rapidly refreshing graphics.